+++ resources/firstnames/male/firstnames-male-x.php
@@ -1,9 +1,9 @@
 <?php return [
-    0 => 'Xavier',
-    1 => 'Xenos',
-    2 => 'Xerxes',
-    3 => 'Xever',
-    4 => 'Ximenes',
-    5 => 'Ximenez',
-    6 => 'Xymenes',
+    'Xavier',
+    'Xenos',
+    'Xerxes',
+    'Xever',
+    'Ximenes',
+    'Ximenez',
+    'Xymenes',
 ];